+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG: cfbd_json_py
 
+## 0.0.3: The "Coaches" Update.
+
+- Implemented `get_cfbd_coaches_info()`, a function that allows a user to get CFB head coach information from a specified season, timeframe, team, or name.
+- Updated the sample script found in the help guide for `get_cfbd_betting_lines()`.
+- Fixed a bug in `get_cfbd_betting_lines()` that would always result in the function not recognizing if the inputed `api_key` is the placeholder key inside the sample script for this function.
+- Updated the package version to `0.0.3`
+
 ## 0.0.2: The "Betting" Update.
 
 - Removed the option to cache data. This may be implemented in a future version.